requirements.txt
streamlit>=1.35.0
groq>=0.9.0
pyodbc>=5.1.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
pandas>=2.2.0
